+++ 3D-Tests/src/main/java/physics/EnumPlayerMovementType.java
@@ -1,13 +1,13 @@
 package main.java.physics;
 
 public enum EnumPlayerMovementType {
-    LEFT(-1.0d),
-    RIGHT(1.0d),
-    STAND(0.0d);
+    LEFT(-1.0f),
+    RIGHT(1.0f),
+    STAND(0.0f);
 
-    public final double directionFactor;
+    public final float directionFactor;
 
-    EnumPlayerMovementType(double directionFactor){
+    EnumPlayerMovementType(float directionFactor){
         this.directionFactor = directionFactor;
     }
 }
